Step 13.23: Move signIn logic to Auth provider
