+++ context.ts
@@ -1,6 +1,3 @@
 import { ModuleContext } from '@graphql-modules/core';
-import { Response } from 'express';
 
-export type MyContext = {
-  res: Response;
-} & ModuleContext;
+export type MyContext = ModuleContext;
