+++ modules/common/index.ts
@@ -56,9 +56,4 @@ export default new GraphQLModule({
     },
     Database,
   ],
-  async context({ res }) {
-    return {
-      res,
-    };
-  },
 });
